# Make `for iter::CartesianIndices` better vectorized for 1d/2d cases.