@N5N3 2022-06-18
With today's master (LLVM14), this PR also helps vectolize 3d CartesianIndices.